+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.1.0
 
 * Add `packWithByteArray` and `packWithMutableByteArray`
+* Fix infinite loop during list decoding when length encoded was negative
 
 ## 0.1.0.0